+++ src/sim/CombatSimulation.ts
@@ -385,7 +385,7 @@ export class CombatSimulation {
             throw new Error("Simulation not finished yet!"); 
         }
 
-        let challengeXPRewards = this.fortressRoom.computeChallengeXPRewards(this.isWin!); 
+        let challengeXPRewards = this.fortressRoom.computeChallengeXPRewards(this.isWin!, this.params.useSponsoredFortressRewards); 
         let wizardResults: Array<CombatSimulationResultsWizard> = this.wizards.map(wizard => {
             return {
                 playerIndex: wizard.playerIndex,